+++ src/main/java/com/cinchapi/ccl/Parser.java
@@ -25,6 +25,7 @@ import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.ThreadSafe;
 
 import com.cinchapi.ccl.grammar.Expression;
+import com.cinchapi.ccl.grammar.KeySymbol;
 import com.cinchapi.ccl.grammar.PostfixNotationSymbol;
 import com.cinchapi.ccl.grammar.Symbol;
 import com.cinchapi.ccl.syntax.AbstractSyntaxTree;
@@ -81,7 +82,7 @@ public abstract class Parser {
      * The ccl statement being parsed.
      */
     protected final String ccl;
-    
+
     /**
      * The dataset used for location resolution.
      */
@@ -117,6 +118,9 @@ public abstract class Parser {
                     if(symbol instanceof Expression) {
                         keys.add(((Expression) symbol).raw().key());
                     }
+                    else if(symbol instanceof KeySymbol) {
+                        keys.add(((KeySymbol) symbol).key());
+                    }
                 });
                 return Collections.unmodifiableSet(keys);
             }
@@ -129,7 +133,6 @@ public abstract class Parser {
      * of symbols in {@link PostfixNotationSymbol postfix notation} that are
      * sorted by the proper order of operations.
      * 
-     * @param symbols a sequential list of tokens
      * @return a {@link Queue} of {@link PostfixNotationSymbol
      *         PostfixNotationSymbols}
      */